% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akseposisjon}
\alias{akseposisjon}
\title{Akseposisjon (venstre-høyre) for de ti største partiene}
\format{
`akseposisjon`:
Et datasett med 10 rader og 2 kolonner:
\describe{
  \item{parti}{Partinavn}
  \item{aksepos}{posisjonen som er forbundet med det gitte partiet}
}
}
\source{
<https://www.stortinget.no/no/Stortinget-og-demokratiet/Organene/Partigruppene/>
}
\usage{
akseposisjon
}
\description{
Akseposisjon (venstre-høyre) for de ti største partiene
}
\keyword{datasets}
