% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mandater}
\alias{mandater}
\title{Mandatfordeling ved valgene i 2017 og 2021}
\format{
`mandater`:
Et datasett med 21 rader og 3 kolonner:
\describe{
  \item{parti}{Partinavn}
  \item{aar}{Valgår}
  \item{mandat}{Antall mandater}
}
}
\source{
<https://www.ssb.no/valg/faktaside/valg>
}
\usage{
mandater
}
\description{
Mandatfordeling ved valgene i 2017 og 2021
}
\keyword{datasets}
