% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partifarger}
\alias{partifarger}
\title{Farger for de ti største partiene}
\format{
`partifarger`:
Et datasett med 10 rader og 2 kolonner:
\describe{
  \item{parti}{Partinavn}
  \item{farge}{Fargen som er forbundet med det gitte partiet}
}
}
\source{
<https://www.stortinget.no/no/Stortinget-og-demokratiet/Organene/Partigruppene/>
}
\usage{
partifarger
}
\description{
Farger for de ti største partiene
}
\keyword{datasets}
