% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_valgoppslutning}
\alias{plot_valgoppslutning}
\title{Plot valgoppslutning}
\usage{
plot_valgoppslutning(aarstall)
}
\arguments{
\item{aarstall}{En character vector av 책r.}
}
\value{
`ggplot2` object, et plot av valgoppslutning for det gitte 책ret (eller 책rene, hvis flere 책r er oppgitt).
}
\description{
Plot valgoppslutning
}
\examples{
plot_valgoppslutning("2017")


}
