% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{velgeroppslutning}
\alias{velgeroppslutning}
\title{Velgeroppsluting ved valgene i 2013, 2017 og 2021}
\format{
`velgeroppslutning`:
Et datasett med 30 rader og 3 kolonner:
\describe{
  \item{parti}{Partinavn}
  \item{aar}{Valgår}
  \item{prosent}{Oppslutningsprosent}
}
}
\source{
<https://www.ssb.no/valg/faktaside/valg>
}
\usage{
velgeroppslutning
}
\description{
Velgeroppsluting ved valgene i 2013, 2017 og 2021
}
\keyword{datasets}
